+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+- StaticTable
+  - Add column clickable
+  - Add active on hover of edit icon
 
 ## 2.0.10